ASYNC-76: allow exceptions on 'thread' to propagate
Similar to 3ea853e.

The thread pool for 'thread' and 'thread-call' catches all exceptions
and prints them. As a result, the exception can never be handled by
the JVM default uncaught exception handler, and the channel returned
from 'thread' is never closed.

This change forces 'thread' to close the channel when there is an
exception and allows the exception to propagate to the default
uncaught exception handler.

Signed-off-by: puredanger <alex.miller@cognitect.com>